feat: extract editing/styling helpers
